src/test/javascript/spec/app/entities/authentic-key/authentic-key.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { AuthenticKeyService } from 'app/entities/authentic-key/authentic-key.service';
import { IAuthenticKey, AuthenticKey } from 'app/shared/model/authentic-key.model';

describe('Service Tests', () => {
    describe('AuthenticKey Service', () => {
        let injector: TestBed;
        let service: AuthenticKeyService;
        let httpMock: HttpTestingController;
        let elemDefault: IAuthenticKey;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(AuthenticKeyService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new AuthenticKey(0, 'AAAAAAA', false, false, currentDate);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        validationdate: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a AuthenticKey', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        validationdate: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        validationdate: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new AuthenticKey(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a AuthenticKey', async () => {
                const returnedFromService = Object.assign(
                    {
                        uniqueKey: 'BBBBBB',
                        assignmentStatus: true,
                        validStatus: true,
                        validationdate: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        validationdate: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of AuthenticKey', async () => {
                const returnedFromService = Object.assign(
                    {
                        uniqueKey: 'BBBBBB',
                        assignmentStatus: true,
                        validStatus: true,
                        validationdate: currentDate.format(DATE_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        validationdate: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a AuthenticKey', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
